made Msg trait global
