+++ modules/src/ics03_connection/msgs.rs
@@ -1,10 +1,10 @@
 use crate::ics03_connection::connection::Counterparty;
 use crate::ics03_connection::error::Kind;
 use crate::ics03_connection::exported::ConnectionCounterparty;
-use crate::ics04_channel::msgs::Msg;
 use crate::ics23_commitment::CommitmentPrefix;
 use crate::ics24_host::identifier::{ClientId, ConnectionId};
 use crate::ics24_host::validate::{validate_client_identifier, validate_connection_identifier};
+use crate::tx_msg::Msg;
 use serde_derive::{Deserialize, Serialize};
 use tendermint::account::Id as AccountId;
