+++ modules/src/lib.rs
@@ -29,6 +29,7 @@ pub mod ics24_host;
 pub mod keys;
 pub mod path;
 pub mod query;
+pub mod tx_msg;
 
 /// Height of a block, same as in `tendermint` crate
 pub type Height = tendermint::lite::Height;
